LayoutTests/http/tests/inspector-protocol/access-inspected-object-expected.txt

CONSOLE MESSAGE: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a frame with origin "http://localhost:8000". Protocols, domains, and ports must match.
Test that code evaluated in the main frame cannot access $0 that resolves into a node in a frame from a different domain. Bug 105423.

 
